+++ app/controllers/action_hooks/webhook_controller.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module ActionHooks
+  class WebhookController < ActionController::API
+    include ActionHooks::WebhookControllerBehavior
+
+    def create
+      process_webhook(@webhook_request)
+      head :ok
+    end
+
+    private
+
+    def process_webhook(webhook_request)
+      worker_class = webhook_source_config.worker
+      worker_class&.constantize&.perform_later(webhook_request.id)
+    end
+  end
+end